Add redirect rule for /datasets to /view
Closes #8

modifies both the httpd.conf and httpd-ssl.conf files to include
redirects from a persistent URL /datasets to the URL that is currently
used by 2.16.0 version of Metacat. A persistent URL is used to prevent
any dead links in the citation.

Add modified files

These files are what create the patch files